Update to add PlotChunkGenerator and remove Fabric18Text.
This update removes the Fabric18Text class. It also includes a new file, PlotChunkGenerator.java, that configures plots with sizes, road widths, and edge widths. Other modified files involve improvements in code reformatting, adjustment of certain argument conditions, updated world settings for the in-game /multiworld command, and adjust gradle dependencies.

+++ common/src/main/java/me/isaiah/multiworld/command/SpawnCommand.java
@@ -60,7 +60,7 @@ public class SpawnCommand {
 		WorldProperties prop = world.getLevelProperties();
         BlockPos pos = new BlockPos(prop.getSpawnX(), prop.getSpawnY(), prop.getSpawnZ());
         if (!world.getWorldBorder().contains(pos)) {
-            pos = world.getTopPosition(Heightmap.Type.MOTION_BLOCKING, new BlockPos(world.getWorldBorder().getCenterX(), 0.0, world.getWorldBorder().getCenterZ()));
+            pos = world.getTopPosition(Heightmap.Type.MOTION_BLOCKING, new BlockPos((int) world.getWorldBorder().getCenterX(), (int) 0.0, (int) world.getWorldBorder().getCenterZ()));
         }
         return pos;
     }
